requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
openai
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
chromadb
sentence-transformers
numpy